requirements.txt
requests~=2.26.0
torch~=1.10.2
flask~=2.0.2
werkzeug~=2.0.2
numpy~=1.21.2
pillow~=8.4.0
torchvision~=0.11.3
tensorflow~=2.7.0
tflite-support~=0.3.1
onnx~=1.10.2
onnxruntime~=1.10.0
onnx-simplifier~=0.3.6
onnx-tf~=1.9.0